Uploading Database and its Operations
This is the final version of the database and its functions.

It connects to the database hosted on my local machine and runs in a virtual environment.
Only the final is uploaded as I will not upload the virtual machine; previous versions were made in a separate file.

ddfunctions.py contains the CRUD operations.